♻️ refactor:  configフォルダの廃止
#78
